all player movement is automatic now
all player movement except for the positioning below the tower is
automatic now.
The questions are presented automatically when the player moves in a
direction.
The question presented is based on the tower color and whether the
automatic dice roll is even or uneven. Turns are automatically ended
after movement is complete. When you try to move right when at the
uttermost right position you are now placed at the uttermost left
tower(to simulate the roundness of the tower) vice-versa for the other
side.